make 'data' field of struct kevent 64bit change briefly discussed on bsd-api-discuss list

+++ sys/sys/event.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: event.h,v 1.1.1.1.2.9 2002/09/29 09:29:26 jdolecek Exp $	*/
+/*	$NetBSD: event.h,v 1.1.1.1.2.10 2002/10/02 19:07:05 jdolecek Exp $	*/
 /*-
  * Copyright (c) 1999,2000,2001 Jonathan Lemon <jlemon@FreeBSD.org>
  * All rights reserved.
@@ -56,7 +56,7 @@ struct kevent {
 	uint32_t	filter;		/* filter for event */
 	uint32_t	flags;		/* action flags for kqueue */
 	uint32_t	fflags;		/* filter flag value */
-	intptr_t	data;		/* filter data value */
+	int64_t		data;		/* filter data value */
 	void		*udata;		/* opaque user data identifier */
 };
